# [F#] Try to speedup a little