Switch "posts" to Span instead of List (ResizeArray), seems to have better read access
Also try to fix bad IL gen for  `taggedPostCount[otherPostIdx] <- taggedPostCount[otherPostIdx] + 1uy`
Dunno if it will be actually faster on test machine.

Old:
```
          // [46 17 - 46 46]
          IL_0045: ldarga.s     taggedPostCount
          IL_0047: ldloc.s      otherPostIdx
          IL_0049: call         instance !0/*unsigned int8*/& valuetype [System.Runtime]System.Span`1<unsigned int8>::get_Item(int32)
          IL_004e: stloc.s      V_8
          IL_0050: ldloc.s      V_8
          IL_0052: ldarga.s     taggedPostCount
          IL_0054: ldloc.s      otherPostIdx
          IL_0056: call         instance !0/*unsigned int8*/& valuetype [System.Runtime]System.Span`1<unsigned int8>::get_Item(int32)
          IL_005b: stloc.s      V_10
          IL_005d: ldloc.s      V_10
          IL_005f: ldobj        [System.Runtime]System.Byte
          IL_0064: stloc.s      V_9
          IL_0066: ldc.i4.1
          IL_0067: stloc.s      V_11
          IL_0069: ldloc.s      V_9
          IL_006b: ldloc.s      V_11
          IL_006d: add
          IL_006e: conv.u1
          IL_006f: stobj        [System.Runtime]System.Byte
          IL_0074: ldloc.s      V_6
          IL_0076: ldc.i4.1
          IL_0077: add
          IL_0078: stloc.s      V_6
```


New:
```
          // [41 17 - 41 52]
          IL_0050: ldarga.s     taggedPostCount
          IL_0052: ldloc.s      otherPostIdx
          IL_0054: call         instance !0/*unsigned int8*/& valuetype [System.Runtime]System.Span`1<unsigned int8>::get_Item(int32)
          IL_0059: stloc.s      V_9
          IL_005b: ldloc.s      V_9
          IL_005d: ldloc.s      V_9
          IL_005f: ldobj        [System.Runtime]System.Byte
          IL_0064: stloc.s      V_10
          IL_0066: ldc.i4.1
          IL_0067: stloc.s      V_11
          IL_0069: ldloc.s      V_10
          IL_006b: ldloc.s      V_11
          IL_006d: add
          IL_006e: conv.u1
          IL_006f: stobj        [System.Runtime]System.Byte
          IL_0074: ldloc.s      V_7
          IL_0076: ldc.i4.1
          IL_0077: add
          IL_0078: stloc.s      V_7
          ```