@jinyus 2023-11-08
Main branch:
```
F# (AOT) | 27.46 ms | 389.67 ms | 3.42 s
F# (JIT) | 91.54 ms | 529.33 ms | 4.05 s
```

This PR:

Ran twice with the same results. Faster for AOT and slower overall for JIT

```
F# (AOT):

        Benchmark 1: ./bin/release/net8.0/aot/fsharp
        Processing time (w/o IO): 26.6695ms
        Processing time (w/o IO): 26.4481ms
        Processing time (w/o IO): 26.5587ms
        Processing time (w/o IO): 26.5295ms
        Processing time (w/o IO): 26.7847ms
        Processing time (w/o IO): 27.0636ms
        Processing time (w/o IO): 26.5322ms
        Processing time (w/o IO): 26.6891ms
        Processing time (w/o IO): 26.5077ms
        Processing time (w/o IO): 26.7149ms
        Processing time (w/o IO): 26.441ms
        Processing time (w/o IO): 26.715ms
        Processing time (w/o IO): 27.0848ms
          Time (mean ± σ):     733.7 ms ±   4.7 ms    [User: 715.6 ms, System: 76.5 ms]
          Range (min … max):   729.1 ms … 744.9 ms    10 runs
         
F# (AOT):

        Benchmark 1: ./bin/release/net8.0/aot/fsharp
        Processing time (w/o IO): 371.4128ms
        Processing time (w/o IO): 370.2145ms
        Processing time (w/o IO): 370.8578ms
          Time (mean ± σ):      3.415 s ±  0.004 s    [User: 3.372 s, System: 0.234 s]
          Range (min … max):    3.413 s …  3.418 s    2 runs
         
F# (AOT):

        Benchmark 1: ./bin/release/net8.0/aot/fsharp
        Processing time (w/o IO): 3227.299ms
        Processing time (w/o IO): 3222.8747ms
        Processing time (w/o IO): 3231.9855ms
          Time (mean ± σ):     14.704 s ±  0.069 s    [User: 16.223 s, System: 0.611 s]
          Range (min … max):   14.655 s … 14.753 s    2 runs
```
```
F# (JIT):

        Benchmark 1: ./bin/release/net8.0/fsharp
        Processing time (w/o IO): 84.6331ms
        Processing time (w/o IO): 84.9935ms
        Processing time (w/o IO): 85.1834ms
        Processing time (w/o IO): 84.8272ms
        Processing time (w/o IO): 84.7734ms
        Processing time (w/o IO): 86.7572ms
        Processing time (w/o IO): 85.807ms
        Processing time (w/o IO): 84.9073ms
        Processing time (w/o IO): 84.9766ms
        Processing time (w/o IO): 84.9694ms
        Processing time (w/o IO): 85.4264ms
        Processing time (w/o IO): 85.1881ms
        Processing time (w/o IO): 85.3512ms
          Time (mean ± σ):      1.025 s ±  0.006 s    [User: 1.133 s, System: 0.097 s]
          Range (min … max):    1.017 s …  1.034 s    10 runs
         
F# (JIT):

        Benchmark 1: ./bin/release/net8.0/fsharp
        Processing time (w/o IO): 616.1436ms
        Processing time (w/o IO): 556.7162ms
        Processing time (w/o IO): 613.3219ms
          Time (mean ± σ):      3.323 s ±  0.085 s    [User: 3.328 s, System: 0.255 s]
          Range (min … max):    3.263 s …  3.383 s    2 runs
         
F# (JIT):

        Benchmark 1: ./bin/release/net8.0/fsharp
        Processing time (w/o IO): 6432.2982ms
        Processing time (w/o IO): 4839.4378ms
        Processing time (w/o IO): 4847.1213ms
          Time (mean ± σ):     14.994 s ±  0.056 s    [User: 15.920 s, System: 0.542 s]
          Range (min … max):   14.954 s … 15.033 s    2 runs
```