@delneg 2023-11-08
LGTM, didn't try to optimize for JIT frankly.